fix: Remove phpcpd from all projects
It's been failing for month as our version isn't working with Xdebug 3 and we've been ignore that failure ever since.
Removing this unclutters the build output a bit. Should we want to reactive this we can see if there is a new phpcpd version and test against that

PHPStorm also provides this functionality locally